+++ example/a.ts
@@ -1,9 +1,13 @@
 import { Elysia, t } from '../src'
+import { req } from '../test/utils'
 
-new Elysia()
-	.onStart(({ server }) => {
-		console.log(`${server?.url}:${server?.port}`)
-	})
-	.listen(1234, (a) => {
+const app = new Elysia().get('/', 'Static Content', {
+	beforeHandle() {
+		return 'beforeHandle'
+	}
+})
+.listen(3000)
 
-	})
+const response = await app.handle(req('/')).then((x) => x.text())
+
+console.log(app.router.static.http.handlers[0])