+++ src/handler.ts
@@ -7,6 +7,7 @@ import { Cookie } from './cookies'
 import { ELYSIA_RESPONSE } from './error'
 
 import type { Context } from './context'
+import { LocalHook } from './types'
 
 const hasHeaderShorthand = 'toJSON' in new Headers()
 
@@ -1160,3 +1161,23 @@ export const errorToResponse = (error: Error, set?: Context['set']) =>
 			headers: set?.headers
 		}
 	)
+
+export const createStaticHandler = (
+	handle: unknown,
+	hooks: LocalHook<any, any, any, any, any, any, any>,
+	setHeaders: Context['set']['headers'] = {}
+): (() => Response) | undefined => {
+	if (typeof handle === 'function') return
+
+	const response = mapResponse(handle, {
+		headers: setHeaders
+	})
+
+	if (
+		hooks.parse.length === 0 &&
+		hooks.transform.length === 0 &&
+		hooks.beforeHandle.length === 0 &&
+		hooks.afterHandle.length === 0
+	)
+		return response.clone.bind(response)
+}